Rework method generation and method parsing XML doc, added new package "Factory" and "ModelDictionary"
