+++ src/main/java/Factory/DictionaryGenerator.java
@@ -1,0 +1,7 @@
+package Factory;
+
+import ModelDictionary.Dictionary;
+
+public interface DictionaryGenerator { // Интерфейс рандомной генерации
+    public Dictionary DictionaryGenerate();
+}
