+++ src/main/java/MainClass.java
@@ -1,16 +1,19 @@
-import Task1.Dictionary;
-import Task1.DictionaryCreationFactory;
-import Task1.Word;
+import ModelDictionary.Dictionary;
+import Factory.DictionaryCreationFactory;
 
 public class MainClass {
     public static void main(String[] args) {
         Dictionary test = DictionaryCreationFactory.getInstance().getDictionary(false);
-        for (Word word:
-             test.getWords()) {
-            System.out.println(word.getNativeWord());
-            System.out.println(word.getTranscription());
-            System.out.println(word.getTranslation());
+        /*for(int i = 0; i < test.getWords().size(); i++){
+            System.out.println(test.getWords().get(i).getNativeWord());
+            System.out.println(test.getWords().get(i).getTranscription());
+            System.out.println(test.getWords().get(i).getTranslation());
             System.out.println();
-        }
+        }*/
+        System.out.println(test.getUser().getID());
+        System.out.println(test.getUser().getFirstName());
+        System.out.println(test.getUser().getSurname());
+        System.out.println(test.getUser().getLogin());
+        System.out.println(test.getUser().getPassword());
     }
 }
