+++ src/main/java/ModelDictionary/LanguageType.java
@@ -1,4 +1,4 @@
-package Task1;
+package ModelDictionary;
 
 public enum LanguageType {
     ENGLISH,
