--- src/main/java/Task1/DictionaryGenerator.java
@@ -1,5 +1,0 @@
-package Task1;
-
-public interface DictionaryGenerator {
-    public Dictionary generate();
-}
